brush: refactor to keep files of components on PascalCase
