+++ brush/src/components/Map/Point/index.js
